Fixed issue with loading images locking up the UI thread
Binary conversion between an InputStream and a Drawable was being done
in the Handler previously. This meant that the work was being performed
in the UI thread and was locking up the screen when large images were
being loaded. The BinaryFetchThread is now capable of returning a
Drawable instead of an InputStream so the work happens in a background
thread.